Reorganize tests in src/context
