+++ src/context/mod.rs
@@ -146,8 +146,6 @@ mod tests {
     use super::*;
     use crate::node::AudioNode;
 
-    use float_eq::assert_float_eq;
-
     fn require_send_sync_static<T: Send + Sync + 'static>(_: T) {}
 
     #[test]
@@ -156,7 +154,7 @@ mod tests {
         let registration = context.mock_registration();
 
         // we want to be able to ship AudioNodes to another thread, so the Registration should be
-        // Send Sync and 'static
+        // Send, Sync and 'static
         require_send_sync_static(registration);
     }
 
@@ -167,62 +165,17 @@ mod tests {
     }
 
     #[test]
-    fn test_sample_rate_length() {
-        let context = OfflineAudioContext::new(1, 48000, 96000.);
-        assert_float_eq!(context.sample_rate(), 96000., abs_all <= 0.);
-        assert_eq!(context.length(), 48000);
+    fn test_online_audio_context_send_sync() {
+        let options = AudioContextOptions {
+            sink_id: "none".into(),
+            ..AudioContextOptions::default()
+        };
+        let context = AudioContext::new(options);
+        require_send_sync_static(context);
     }
 
     #[test]
-    fn test_decode_audio_data() {
-        let context = OfflineAudioContext::new(1, 1, 44100.);
-        let file = std::fs::File::open("samples/sample.wav").unwrap();
-        let audio_buffer = context.decode_audio_data_sync(file).unwrap();
-
-        assert_eq!(audio_buffer.sample_rate(), 44100.);
-        assert_eq!(audio_buffer.length(), 142_187);
-        assert_eq!(audio_buffer.number_of_channels(), 2);
-        assert_float_eq!(audio_buffer.duration(), 3.224, abs_all <= 0.001);
-
-        let left_start = &audio_buffer.get_channel_data(0)[0..100];
-        let right_start = &audio_buffer.get_channel_data(1)[0..100];
-        // assert distinct two channel data
-        assert!(left_start != right_start);
-    }
-
-    // #[test]
-    // disabled: symphonia cannot handle empty WAV-files
-    #[allow(dead_code)]
-    fn test_decode_audio_data_empty() {
-        let context = OfflineAudioContext::new(1, 1, 44100.);
-        let file = std::fs::File::open("samples/empty_2c.wav").unwrap();
-        let audio_buffer = context.decode_audio_data_sync(file).unwrap();
-        assert_eq!(audio_buffer.length(), 0);
-    }
-
-    #[test]
-    fn test_decode_audio_data_decoding_error() {
-        let context = OfflineAudioContext::new(1, 1, 44100.);
-        let file = std::fs::File::open("samples/corrupt.wav").unwrap();
-        assert!(context.decode_audio_data_sync(file).is_err());
-    }
-
-    #[test]
-    fn test_create_buffer() {
-        let number_of_channels = 3;
-        let length = 2000;
-        let sample_rate = 96_000.;
-
-        let context = OfflineAudioContext::new(1, 1, 44100.);
-        let buffer = context.create_buffer(number_of_channels, length, sample_rate);
-
-        assert_eq!(buffer.number_of_channels(), 3);
-        assert_eq!(buffer.length(), 2000);
-        assert_float_eq!(buffer.sample_rate(), 96000., abs_all <= 0.);
-    }
-
-    #[test]
-    fn test_registration() {
+    fn test_context_equals() {
         let context = OfflineAudioContext::new(1, 48000, 96000.);
         let dest = context.destination();
         assert!(dest.context() == context.base());
